% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{build_kernelMatrix}
\alias{build_kernelMatrix}
\title{Build the Gaussian kernel matrix}
\usage{
build_kernelMatrix(object, method = NULL, bw = NULL)
}
\arguments{
\item{object}{The Tessa object.}

\item{method}{The method of bandwidth selection, default is NULL}
}
\value{
The Tessa object with the Gaussian kernel matrix.
}
\description{
Build the Gaussian kernel matrix
}
