% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{build_kernelMatrix_lineage}
\alias{build_kernelMatrix_lineage}
\title{Build the Gaussian kernel matrix for one lineage}
\usage{
build_kernelMatrix_lineage(object, lineage = "lineage1", bw)
}
\arguments{
\item{object}{The Tessa object.}

\item{lineage}{The name of the lineage to use}

\item{bw}{The numeric bandwidth}
}
\value{
The Tessa object with the Gaussian kernel matrix.
}
\description{
Build the Gaussian kernel matrix for one lineage
}
