% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{CreateTessaObject}
\alias{CreateTessaObject}
\title{Create the Tessa object}
\usage{
CreateTessaObject(
  counts,
  normalized = FALSE,
  meta_df,
  covariates = NULL,
  signature_genes = NULL
)
}
\arguments{
\item{counts}{The gene expression matrix of dimension n x G,
with genes in rows and spots in columns.}

\item{normalized}{If TRUE, gene_expression is normalized. Otherwise, gene_expression is raw count}

\item{meta_df}{The n x 3 dimensional dataframe contains location coordinates and pseudotime.
The rownames of meta_df should match the colnames of counts matrix.
The colnames of meta_df should be x, y and t}

\item{covariates}{The covariate (if any) design matrix of dimension n x p,
modeling gene expression.
The rownames of covariates matrix should match the colnames of counts matrix.}
}
\value{
returns Tessa object.
}
\description{
Create the Tessa object
}
