% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{data_preprocess}
\alias{data_preprocess}
\title{Preprocess: normalize count matrix, filter out genes and spots}
\usage{
data_preprocess(
  object,
  spot.threshold = 10,
  gene.threshold = 0.1,
  normalized = NULL
)
}
\arguments{
\item{object}{Tessa object}

\item{spot.threshold}{A number to filter out spots with less than spot.threshold gene counts}

\item{gene.threshold}{A number to filter out genes expressed less than gene.threshold percent of spots}

\item{normalized}{If TRUE, object@gene_expression is normalized, skip normalization step.}
}
\value{
Returns Tessa object
}
\description{
Preprocess: normalize count matrix, filter out genes and spots
}
