% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{gaussian_distance_normalized_kernel}
\alias{gaussian_distance_normalized_kernel}
\title{Build the Gaussian kernel matrix with normalized distance matrix}
\usage{
gaussian_distance_normalized_kernel(X = NULL, sigma = NULL)
}
\arguments{
\item{X}{N by D numeric data matrix}

\item{sigma}{Positive scalar that specifies the bandwidth of the Gaussian kernel}
}
\description{
Build the Gaussian kernel matrix with normalized distance matrix
}
