% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{get_Test1_result}
\alias{get_Test1_result}
\title{Retrieve Test1 result for one or all lineages}
\usage{
get_Test1_result(object, lineage = NULL)
}
\arguments{
\item{lineage}{Name of lineage}

\item{obejct}{A TESSA object}
}
\value{
A dataframe of Test1 result
}
\description{
Retrieve Test1 result for one or all lineages
}
