% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{get_Test1_topK_genes}
\alias{get_Test1_topK_genes}
\title{Retrieve topK uTSVGs for each lineage}
\usage{
get_Test1_topK_genes(Tessa.obj, topK = NULL)
}
\arguments{
\item{Tessa.obj}{A TESSA object}

\item{topK}{If NULL, return all uTSVGs; otherwise, return topK uTSVGs of each lineage}
}
\value{
A list of uTSVGs of each lineage
}
\description{
Retrieve topK uTSVGs for each lineage
}
