% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{invert}
\alias{invert}
\title{Inverse of a matrix using RcppArmadillo}
\usage{
invert(X)
}
\arguments{
\item{X}{A matrix}
}
\value{
The inverse matrix
}
\description{
Inverse of a matrix using RcppArmadillo
}
