% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{legacy_Test2}
\alias{legacy_Test2}
\title{Legacy Individual Test for TVG and SVG}
\usage{
legacy_Test2(object, gene, K_test, lineage = "lineage1")
}
\arguments{
\item{object}{A TESSA object}

\item{gene}{Gene to test}

\item{K_test}{Name of the kernel being tested}

\item{lineage}{Name of the lineage}
}
\value{
A list containing the gene, variance components, score, degrees of
freedom, scale, and p-value
}
\description{
Original Test 2 implementation retained for reproducibility.
}
