% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normal the matrix by min max value}
\usage{
normalize_matrix(M)
}
\arguments{
\item{M}{numeric data matrix}
}
\description{
Normal the matrix by min max value
}
