% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{run_Test1}
\alias{run_Test1}
\title{Run Test1 for all genes, all lineages}
\usage{
run_Test1(object, LOO = FALSE, LOO_pv_threshold = 0.05, npcs = 10, acc = 1e-15)
}
\arguments{
\item{object}{The TESSA object}

\item{LOO}{If TRUE, do LOO double dipping correction}

\item{npcs}{The PC number to use in slingshot pseudotime estimation}

\item{acc}{If pvalue has too many zeros, decrease this number to adjust the lower limit of pvalue}
}
\value{
The TESSA object
}
\description{
Run Test1 for all genes, all lineages
}
