% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{run_Test1_lineage}
\alias{run_Test1_lineage}
\title{Run Test1 for all genes}
\usage{
run_Test1_lineage(object, lineage = "lineage1", acc = 1e-07)
}
\arguments{
\item{object}{The TESSA object}

\item{lineage}{Name of the lineage}

\item{acc}{If pvalue has too many zeros, decrease this number to adjust the lower limit of pvalue}
}
\value{
The TESSA object
}
\description{
Run Test1 for all genes
}
