% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{run_Test1_lineage_LOO_pergene}
\alias{run_Test1_lineage_LOO_pergene}
\title{Run Test1 for one gene, one lineage}
\usage{
run_Test1_lineage_LOO_pergene(
  object,
  gene,
  lineage = "lineage1",
  acc = 1e-07,
  npcs = 30
)
}
\arguments{
\item{object}{The TESSA object}

\item{gene}{The gene to do Test1}

\item{lineage}{The name of lineage}
}
\value{
pvalue of test1
}
\description{
Run Test1 for one gene, one lineage
}
