% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{run_Test2_lineage}
\alias{run_Test2_lineage}
\title{Run Test2 for selected genes for one lineage}
\usage{
run_Test2_lineage(
  object,
  lineage = "lineage1",
  genes = NULL,
  LOO = FALSE,
  npcs = 30,
  parallel = TRUE,
  workers = 2L,
  mc.preschedule = TRUE,
  control_blas = TRUE
)
}
\arguments{
\item{object}{The TESSA object}

\item{lineage}{Name of the lineage to test}

\item{genes}{The genes to do Test2. If NULL, then test all the uTSVGs}

\item{LOO}{If TRUE, run LOO double dipping correction for Test2}

\item{npcs}{The PC number to use in slingshot pseudotime estimation}

\item{parallel}{If TRUE, run independent gene tests with fork-based parallelism on Unix-like systems; Windows falls back to serial execution}

\item{workers}{Number of fork workers used when parallel is TRUE}

\item{mc.preschedule}{Passed to parallel::mclapply; FALSE is slower but isolates gene-level errors more clearly}

\item{control_blas}{If TRUE, use RhpcBLASctl when available to limit BLAS and OpenMP to one thread per fork worker}
}
\description{
Run Test2 for selected genes for one lineage
}
