% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\docType{class}
\name{Tessa-class}
\alias{Tessa-class}
\title{Each Tessa object has a number of slots which store information. Key slots to access
are listed below.}
\description{
Each Tessa object has a number of slots which store information. Key slots to access
are listed below.
}
\section{Slots}{

\describe{
\item{\code{gene_expression}}{The gene expression matrix with genes in rows
and spots in columns.}

\item{\code{meta_df}}{The location coorinates and pseudotime aframe}

\item{\code{covariates}}{The covariate design matrix modeling gene expression.
#! try covariates in}

\item{\code{kernel}}{The kernel matrix to capture spatial correlation, trajectory correlation and error term}

\item{\code{bandwidth}}{The bandwidth parameter of Gaussian kernel.}

\item{\code{normalized}}{If TRUE, gene_expression is normalized. Otherwise, gene_expression is raw count}

\item{\code{signature_genes}}{The vector of gene names, which are used to estimate pseudotime. This is for LOO double dipping correction step}

\item{\code{result}}{The summary of uTSVGs, TVGs and SVGs.}
}}

