% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{Test1}
\alias{Test1}
\title{Overall Test for uTSVG}
\usage{
Test1(Y, covariates = NULL, kernel_mat_list, acc = 1e-07)
}
\arguments{
\item{Y}{The numeric matrix with n spots (row) and k genes (column)}

\item{covariates}{Covariates X as fixed effect}

\item{kernel_mat_list}{A list of kernel matrix, as random effect. kernel_S, kernel_T and kernel_error}
}
\description{
Overall Test for uTSVG
}
