% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{build_kernelMatrix}
\alias{build_kernelMatrix}
\title{Build the Gaussian kernel matrix}
\usage{
build_kernelMatrix(object)
}
\arguments{
\item{object}{The Tessa object.}
}
\value{
The Tessa object with the Gaussian kernel matrix.
}
\description{
Build the Gaussian kernel matrix
}
