% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{run_Test2}
\alias{run_Test2}
\title{Run Test1 for all lineages}
\usage{
run_Test2(object, genes = NULL, pv_threshold = 0.05, LOO = FALSE)
}
\arguments{
\item{pv_threshold}{The pvalue threshold for test1, to get the uTSVGs that are significant in Test1}
}
\description{
Run Test1 for all lineages
}
