% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{run_Test2_lineage}
\alias{run_Test2_lineage}
\title{Run Test2 for selected genes for one lineage}
\usage{
run_Test2_lineage(
  object,
  lineage = "lineage1",
  genes = NULL,
  LOO = FALSE,
  DP_genes = NULL
)
}
\arguments{
\item{genes}{The genes to test for Test2
slingshot'}
}
\description{
Run Test2 for selected genes for one lineage
}
