% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{run_Test2_lineage}
\alias{run_Test2_lineage}
\title{Run Test2 for selected genes for one lineage}
\usage{
run_Test2_lineage(
  object,
  lineage = "lineage1",
  genes = NULL,
  LOO = FALSE,
  npcs = 30,
  num_cores = 1,
  parallel = FALSE
)
}
\arguments{
\item{genes}{The genes to do Test2. If NULL, then test all the uTSVGs}

\item{LOO}{If TRUE, run LOO double dipping correction for Test2}

\item{npcs}{The PC number to use in slingshot pseudotime estimation}

\item{obejct}{The TESSA object}

\item{pv_threshold}{The pvalue threshold for test1, to get the uTSVGs that are significant in Test1}
}
\description{
Run Test2 for selected genes for one lineage
}
