% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{Test2}
\alias{Test2}
\title{Individual Test for TVG and SVG}
\usage{
Test2(object, gene, K_test, lineage = "lineage1", num_cores = 1)
}
\arguments{
\item{Y}{The numeric matrix with n spots (row) and k genes (column)}

\item{covariates}{Covariates X as fixed effect}

\item{kernel_mat_list}{A list of kernel matrix, as random effect. kernel_S, kernel_T and kernel_error}
}
\description{
Individual Test for TVG and SVG
}
