% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESSA.R
\name{TT}
\alias{TT}
\title{Trace of matrix multiplication using RcppArmadillo}
\usage{
TT(M1, M2)
}
\arguments{
\item{M1}{A numeric matrix}

\item{M2}{A numeric matrix}
}
\value{
Trace of matrix multiplication
}
\description{
Trace of matrix multiplication using RcppArmadillo
}
